TG: ODScene - change tonemapping members to double, bind skylight texture for rendering; test.menu - add tonemapping items; projected_grid.effect - add skylight sampler, write some code using it, far from complete and flawless

+++ AppsV2/OceanDemo/ODScene.h
@@ -50,11 +50,11 @@
     BOOL disconnecting;
 
     // tonemapping parameters
-    float referenceWhite;
-    float key;
-    float adaptationTimeScale;
-    float lastFrameLuminance;
-    float currentFrameLuminance;
+    double referenceWhite;
+    double key;
+    double adaptationTimeScale;
+    double lastFrameLuminance;
+    double currentFrameLuminance;
 
     //
     IVector2 currentResolution;
